backend/requirements.txt

Django>=5.2
djangorestframework>=3.15
drf-yasg>=1.21
python-dotenv>=1.0.0
